chore: relocate canonical metadata and dedupe group definitions
