+++ analysis/metadata/run2_scenarios.yaml
@@ -1,10 +1,8 @@
 # Run 2 scenario composition. This file is the single source of truth
 # for mapping high-level scenario names to channel groups. Each group
-# is defined either in the canonical topeft/params/metadata.yml bundle or the
-# supplemental analysis/metadata/metadata_run2_groups.yaml file for derived
-# layouts (off-Z splits, forward combinations, etc.). Legacy metadata_*.yaml
-# copies are retained only for historical reference.
-# Consumed by topeft.modules.run2_scenarios and the Run-2
+# is defined in the canonical analysis/metadata/metadata.yml bundle.
+# Legacy metadata_*.yaml copies are retained only for historical reference.
+# Consumed by topeft.modules.scenario_groups and the Run-2
 # workflow/datacard tools.
 scenarios:
   TOP_22_006:
