+++ analysis/topeft_run2/configs/fullR2_run.yml
@@ -13,7 +13,7 @@ defaults:
     - ctG
 
   # Select the metadata bundle that drives region and variable planning.
-  metadata: ../../topeft/params/metadata.yml
+  metadata: ../metadata/metadata.yml
 
   # Default scenario grouping used for reinterpretation studies.
   scenarios:
